대입문 시행 과정 임의로 작성함 (심볼테이블 관련 제외)

+++ PJ1/PJ1/parser.hpp
@@ -40,7 +40,7 @@ private:
 	int term();
 	int term_tail();
 	int factor();
-	int factor_tail();
+	double factor_tail();
 
 	int ident();
 	int ident_val();
